requirements.txt
asgiref==3.11.0
certifi==2026.1.4
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
dataclasses==0.6
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djongo==1.2.31
dnspython==2.8.0
idna==3.11
msgpack==1.1.2
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pymongo==4.16.0
python-decouple==3.8
razorpay==2.0.0
redis==7.1.0
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
